requirements.txt
streamlit==1.38.0
streamlit-folium==0.17.0
earthengine-api>=1.0.0
geemap==0.36.0
google-genai
pandas
numpy
scikit-learn

rasterio
folium
plotly==5.24.1
torch
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cpu  # Lite ML
requests
beautifulsoup4
pyyaml
pytest
black
flake8
mypy  # Pro devops
openpyxl
reportlab  # PDF reports
streamlit-aggrid==1.1.1  # Data tables

